reducing the precision to lower the data transfer
+ reduced the precision of the detected angle and delta angle from 16bits to 8bits
+ currently neglecting the precision to lessen the data over communication channel.

+++ App/inc/CSerialProtocol.h
@@ -55,8 +55,8 @@ public:
 
 	typedef struct
 	{
-		int16_t theta;
-		int16_t delta_theta;
+		int8_t theta;
+		int8_t delta_theta;
 		//uint16_t probability;
 		//uint8_t opinion;
 	} object_detection_block_t;
